# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MeetingBot API",
    "version": "2.3.0",
    "description": "Send bots into Zoom, Google Meet, and Microsoft Teams meetings to record, transcribe, and analyse them with Claude (Anthropic) or Gemini (Google) AI.\n\nMulti-tenant: each external service registers an account and gets its own API key. Pre-fund a credit balance via Stripe (card) or USDC (ERC-20) — credits are deducted automatically per bot run.\n\n**Authentication:** Pass your `sk_live_...` API key as a Bearer token on every request:\n```\nAuthorization: Bearer sk_live_<your-key>\n```\n\n**Business accounts:** Upgrade to a business account to isolate data per end-user. Pass `X-Sub-User: <user-id>` on every request — different sub-users cannot see each other's bots.\n\n**Interactive docs:** Public Swagger UI at `/api/docs` · Admin full schema at `/api/v1/admin/docs` (admin only).\n\n**v2.3.0 changes:** Zero-reload async dashboard, admin rate limits (10/min), webhook replay protection via `X-MeetingBot-Timestamp`, bot queue wake latency near-zero via asyncio.Event, analytics caching (30 s), calendar dedup memory fix."
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    }
  ],
  "security": [
    { "BearerAuth": [] }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key obtained from `POST /api/v1/auth/register` or `POST /api/v1/auth/keys`. Format: `sk_live_<40-char-token>`."
      }
    },
...
  },
  "paths": {
...
    "/api/v1/admin/wallet": {
      "get": {
        "tags": ["Admin"],
        "summary": "[Admin] Get platform USDC wallet",
        "description": "Get the current platform USDC collection wallet address. **Requires admin account.**",
        "operationId": "admin_get_wallet",
        "responses": {
          "200": {
            "description": "Platform wallet",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "wallet_address": { "type": "string", "nullable": true, "description": "Platform USDC collection wallet address (ERC-20). Null if not set." }
                  }
                }
              }
            }
          },
          "403": { "description": "Admin access required" }
        }
      },
      "put": {
        "tags": ["Admin"],
        "summary": "[Admin] Set platform USDC wallet",
        "description": "Set or update the platform USDC collection wallet address. **Requires admin account.**",
        "operationId": "admin_set_wallet",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["wallet_address"],
                "properties": {
                  "wallet_address": { "type": "string", "description": "Ethereum address (0x..., 42 characters).", "example": "0xAbCdEf0123456789AbCdEf0123456789AbCdEf01" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Wallet updated" },
          "403": { "description": "Admin access required" },
          "422": { "description": "Invalid Ethereum address" }
        }
      }
    },
    "/api/v1/admin/rpc-url": {
      "get": {
        "tags": ["Admin"],
        "summary": "[Admin] Check RPC URL status",
        "description": "Check whether a `CRYPTO_RPC_URL` is configured (env var or admin-set). **Requires admin account.**",
        "operationId": "admin_get_rpc_url",
        "responses": {
          "200": {
            "description": "RPC URL status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "rpc_url_set": { "type": "boolean" },
                    "source": { "type": "string", "enum": ["env", "db", "none"] }
                  }
                }
              }
            }
          },
          "403": { "description": "Admin access required" }
        }
      },
      "put": {
        "tags": ["Admin"],
        "summary": "[Admin] Set Ethereum RPC URL",
        "description": "Set the Ethereum RPC URL used by the USDC monitor. Stored in DB (no restart needed). The env var `CRYPTO_RPC_URL` always takes precedence if set. URL is validated via `eth_blockNumber` before saving. **Requires admin account.**",
        "operationId": "admin_set_rpc_url",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["rpc_url"],
                "properties": {
                  "rpc_url": { "type": "string", "description": "Ethereum JSON-RPC endpoint URL.", "example": "https://mainnet.infura.io/v3/YOUR_KEY" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "RPC URL saved" },
          "403": { "description": "Admin access required" },
          "422": { "description": "Invalid or unreachable RPC URL" }
        }
      }
    },
    "/api/v1/admin/config": {
      "get": {
        "tags": ["Admin"],
        "summary": "[Admin] List all platform config",
        "description": "List all platform configuration key-value pairs. **Requires admin account.**",
        "operationId": "admin_get_config",
        "responses": {
          "200": {
            "description": "Platform config",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "configs": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "key": { "type": "string" },
                          "value": { "type": "string" },
                          "updated_at": { "type": "string", "format": "date-time", "nullable": true }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": { "description": "Admin access required" }
        }
      }
    },
    "/api/v1/admin/credit": {
      "post": {
        "tags": ["Admin"],
        "summary": "[Admin] Manually credit account",
        "description": "Manually credit a user account. Use to fix missed USDC deposits that the monitor failed to detect (e.g. user sent from an unregistered wallet). **Requires admin account.**",
        "operationId": "admin_credit",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "amount_usd"],
                "properties": {
                  "email": { "type": "string", "format": "email", "description": "Email of the account to credit." },
                  "amount_usd": { "type": "number", "exclusiveMinimum": 0, "description": "Amount in USD to credit (must be positive)." },
                  "note": { "type": "string", "nullable": true, "description": "Optional admin note recorded in the transaction description." }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account credited",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "account_id": { "type": "string" },
                    "email": { "type": "string" },
                    "credited_usd": { "type": "number" },
                    "new_balance_usd": { "type": "number" }
                  }
                }
              }
            }
          },
          "403": { "description": "Admin access required" },
          "404": { "description": "Account not found" }
        }
      }
    },
    "/api/v1/admin/accounts/{account_id}/set-account-type": {
      "post": {
        "tags": ["Admin"],
        "summary": "[Admin] Set account type",
        "description": "Set the account type (`personal` or `business`) for any user account.\n\n- **personal** — standard single-user account.\n- **business** — multi-tenant mode; the account can use `X-Sub-User` to isolate data per end-user.\n\n**Requires admin account.** Also available as a dropdown in the admin panel UI at `/admin`.",
        "operationId": "admin_set_account_type",
        "parameters": [
          { "name": "account_id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "UUID of the account to update." }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["account_type"],
                "properties": {
                  "account_type": {
                    "type": "string",
                    "enum": ["personal", "business"],
                    "description": "Account type to set.",
                    "example": "business"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account type updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "account_id": { "type": "string" },
                    "email": { "type": "string" },
                    "account_type": { "type": "string", "enum": ["personal", "business"] },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "403": { "description": "Admin access required" },
          "404": { "description": "Account not found" },
          "422": { "description": "account_type must be 'personal' or 'business'" }
        }
      }
    },
    "/api/v1/admin/usdc/unmatched": {
      "get": {
        "tags": ["Admin"],
        "summary": "[Admin] List unmatched USDC transfers",
        "description": "List USDC transfers that arrived at the platform wallet but could not be attributed to any user account (sender wallet not registered).\n\nTo resolve: 1) identify the user by `from_address`, 2) credit their account via `POST /admin/credit`, 3) mark resolved via `POST /admin/usdc/unmatched/{tx_hash}/resolve`.\n\n**Requires admin account.**",
        "operationId": "admin_list_unmatched",
        "parameters": [
          { "name": "resolved", "in": "query", "required": false, "schema": { "type": "boolean" }, "description": "Filter by resolved status. Omit for all, `false` for pending only (default), `true` for resolved only." }
        ],
        "responses": {
          "200": {
            "description": "Unmatched transfers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "transfers": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "tx_hash": { "type": "string" },
                          "from_address": { "type": "string" },
                          "to_address": { "type": "string" },
                          "amount_usdc": { "type": "number" },
                          "block_number": { "type": "integer" },
                          "detected_at": { "type": "string", "format": "date-time" },
                          "resolved": { "type": "boolean" },
                          "resolution_note": { "type": "string", "nullable": true }
                        }
                      }
                    },
                    "total": { "type": "integer" }
                  }
                }
              }
            }
          },
          "403": { "description": "Admin access required" }
        }
      }
    },
    "/api/v1/admin/usdc/unmatched/{tx_hash}/resolve": {
      "post": {
        "tags": ["Admin"],
        "summary": "[Admin] Resolve unmatched USDC transfer",
        "description": "Mark an unmatched USDC transfer as resolved after the account has been manually credited. Call this after `POST /admin/credit`. **Requires admin account.**",
        "operationId": "admin_resolve_unmatched",
        "parameters": [
          { "name": "tx_hash", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Ethereum transaction hash." }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "note": { "type": "string", "nullable": true, "description": "Optional note describing the resolution action taken." }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Transfer marked resolved" },
          "403": { "description": "Admin access required" },
          "404": { "description": "Transaction not found" }
        }
      }
    },
    "/api/v1/admin/usdc/rescan": {
      "post": {
        "tags": ["Admin"],
        "summary": "[Admin] Rescan USDC blockchain from block",
        "description": "Reset the USDC monitor's last-processed block so it rescans from a given block on the next cycle (≤60 s). Duplicate transactions are skipped automatically via idempotency checks. **Requires admin account.**",
        "operationId": "admin_rescan_usdc",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["from_block"],
                "properties": {
                  "from_block": { "type": "integer", "minimum": 0, "description": "Block number to start rescanning from." }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rescan queued",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "from_block": { "type": "integer" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "403": { "description": "Admin access required" }
        }
      }
    }
  },
...
}